Module polars_parquet::arrow::read

APIs to read from Parquet format.

Re-exports

pub use schema::infer_schema;

pub use schema::FileMetaData;

pub use crate::parquet::error::Error as ParquetError;

pub use crate::parquet::metadata::ColumnChunkMetaData;

pub use crate::parquet::metadata::ColumnDescriptor;

pub use crate::parquet::metadata::RowGroupMetaData;

pub use crate::parquet::page::CompressedDataPage;

pub use crate::parquet::page::DataPageHeader;

pub use crate::parquet::page::Page;

pub use crate::parquet::read::decompress;

pub use crate::parquet::read::get_column_iterator;

pub use crate::parquet::read::read_columns_indexes as _read_columns_indexes;

pub use crate::parquet::read::read_metadata as _read_metadata;

pub use crate::parquet::read::read_pages_locations;

pub use crate::parquet::read::BasicDecompressor;

pub use crate::parquet::read::Decompressor;

pub use crate::parquet::read::MutStreamingIterator;

pub use crate::parquet::read::PageFilter;

pub use crate::parquet::read::PageReader;

pub use crate::parquet::read::ReadColumnIterator;

pub use crate::parquet::read::State;

pub use crate::parquet::schema::types::GroupLogicalType;

pub use crate::parquet::schema::types::ParquetType;

pub use crate::parquet::schema::types::PhysicalType;

pub use crate::parquet::schema::types::PrimitiveConvertedType;

pub use crate::parquet::schema::types::PrimitiveLogicalType;

pub use crate::parquet::schema::types::TimeUnit as ParquetTimeUnit;

pub use crate::parquet::types::int96_to_i64_ns;

pub use crate::parquet::fallible_streaming_iterator;

Modules

indexes

API to perform page-level filtering (also known as indexes)

schema

APIs to handle Parquet <-> Arrow schemas.

statistics

APIs exposing crate::parquet’s statistics as arrow’s statistics.

Structs

FileReader

An iterator of Chunks coming from row groups of a parquet file.

NestedState

The state of nested data types.

RowGroupDeserializer

An Iterator of Chunk that (dynamically) adapts a vector of iterators of Array into an iterator of Chunk.

RowGroupReader

An [Iterator<Item=RowGroupDeserializer>] from row groups of a parquet file.

StructIterator

An iterator adapter over NestedArrayIter assumed to be encoded as Struct arrays

Enums

InitNested

The initial info of nested data types.

Traits

FallibleStreamingIterator

A fallible, streaming iterator.

Pages

Trait describing a FallibleStreamingIterator of Page

Functions

column_iter_to_arrays

An iterator adapter that maps multiple iterators of Pages into an iterator of Arrays.

create_list

Creates a new ListArray or FixedSizeListArray.

create_map

Creates a new MapArray.

get_field_columns

Returns all ColumnChunkMetaData associated to field_name. For non-nested parquet types, this returns a single column

get_field_pages

Returns all ColumnChunkMetaData associated to field_name. For non-nested parquet types, this returns a single column

get_page_iterator

Creates a new iterator of compressed pages.

init_nested

Initialize NestedState from &[InitNested].

n_columns

Returns the number of (parquet) columns that a ArrowDataType contains.

read_columns

Reads all columns that are part of the parquet field field_name

read_columns_many

Returns a vector of iterators of Array (ArrayIter) corresponding to the top level parquet fields whose name matches fields’s names.

read_metadata

Reads parquets’ metadata synchronously.

to_deserializer

Converts a vector of columns associated with the parquet field whose name is Field to an iterator of Array, ArrayIter of chunk size chunk_size.

Type Aliases

ArrayIter

Type def for a sharable, boxed dyn Iterator of arrays

NestedArrayIter

Type def for a sharable, boxed dyn Iterator of NestedStates and arrays