polars_parquet::arrow::read::statistics

Function deserialize_all

pub fn deserialize_all(
    field: &Field,
    row_groups: &[RowGroupMetadata],
    field_idx: usize,
) -> ParquetResult<Option<ArrowColumnStatisticsArrays>>

Deserializes the statistics in the column chunks from a single row_group into Statistics associated from field’s name.

Errors

This function errors if the deserialization of the statistics fails (e.g. invalid utf8)