polars_parquet::arrow::write

Struct FileWriter

pub struct FileWriter<W: Write> { /* private fields */ }

An interface to write a parquet to a Write

Implementations

impl<W: Write> FileWriter<W>

pub fn options(&self) -> WriteOptions

The options assigned to the file

pub fn parquet_schema(&self) -> &SchemaDescriptor

The SchemaDescriptor assigned to this file

pub fn schema(&self) -> &ArrowSchema

The ArrowSchema assigned to this file

impl<W: Write> FileWriter<W>

pub fn new_with_parquet_schema( writer: W, schema: ArrowSchema, parquet_schema: SchemaDescriptor, options: WriteOptions, ) -> Self

Returns a new FileWriter.

Error

If it is unable to derive a parquet schema from ArrowSchema.

pub fn try_new( writer: W, schema: ArrowSchema, options: WriteOptions, column_options: &[ColumnWriteOptions], ) -> PolarsResult<Self>

Returns a new FileWriter.

Error

If it is unable to derive a parquet schema from ArrowSchema.

pub fn write( &mut self, row_group: RowGroupIterColumns<'_, PolarsError>, ) -> PolarsResult<()>

Writes a row group to the file.

pub fn end( &mut self, key_value_metadata: Option<Vec<KeyValue>>, column_options: &[ColumnWriteOptions], ) -> PolarsResult<u64>

Writes the footer of the parquet file. Returns the total size of the file. If key_value_metadata is provided, the value is taken as-is. If it is not provided, the Arrow schema is added to the metadata.

pub fn into_inner(self) -> W

Consumes this writer and returns the inner writer

pub fn into_inner_and_metadata(self) -> (W, ThriftFileMetadata)

Returns the underlying writer and ThriftFileMetadata

Panics

This function panics if Self::end has not yet been called