polars_parquet::parquet

Module compression

Functionality to compress and decompress data according to the parquet specification

Structs

BrotliLevel

Represents a valid brotli compression level.

GzipLevel

Represents a valid gzip compression level.

ZstdLevel

Represents a valid zstd compression level.

Enums

Compression

CompressionOptions

Defines the compression settings for writing a parquet file.

DecompressionContext

Functions

compress

Compresses data stored in slice input_buf and writes the compressed result to output_buf.

decompress

Decompresses data stored in slice input_buf and writes output to output_buf. Returns the total number of bytes written.