Module polars_parquet::parquet::deserialize

Structs

FilteredHybridBitmapIter

An Iterator adapter over HybridEncoded that yields FilteredHybridEncoded.

FixexBinaryIter

HybridRleBooleanIter

An iterator adapter that maps an iterator of HybridEncoded into an iterator over bool.

HybridRleIter

An iterator of HybridEncoded, adapter over hybrid_rle::HybridEncoded.

OptionalValues

Iterator adapter to convert an iterator of non-null values and an iterator over validity into an iterator of optional values.

SliceFilteredIter

An iterator adapter that converts an iterator over items into an iterator over slices of those N items.

Enums

BinaryPageState

BooleanPageState

DefLevelsDecoder

Decoder of definition levels.

FilteredHybridEncoded

The decoding state of the hybrid-RLE decoder with a maximum definition level of 1 that can supports skipped runs

FixedLenBinaryPageState

HybridEncoded

The decoding state of the hybrid-RLE decoder with a maximum definition level of 1

NativePageState

The deserialization state of a DataPage of Primitive parquet primitive type

Traits

HybridRleRunsIterator

Functions

native_cast

Views the values of the data page as Casted to NativeType.

Type Aliases

Casted

Typedef of an iterator over PLAIN page values

FilteredHybridRleDecoderIter

Type definition of a FilteredHybridBitmapIter of HybridDecoderBitmapIter.

HybridDecoderBitmapIter

Type definition for a HybridRleIter using hybrid_rle::Decoder.

HybridRleDecoderIter

Type definition for a HybridRleBooleanIter using hybrid_rle::Decoder.