Module polars_parquet::parquet::indexes

Structs

BooleanIndex

An index of a column of boolean physical type

ByteIndex

An index of a column of bytes physical type

FilteredPage

An enum describing a page that was either selected in a filter pushdown or skipped

FixedLenByteIndex

An index of a column of fixed len byte physical type

Interval

An interval

NativeIndex

An index of a column of NativeType physical representation

PageIndex

The index of a page, containing the min and max values of the page.

PageLocation

Enums

BoundaryOrder

Enum to annotate whether lists of min/max elements inside ColumnIndex are ordered and if so, in which direction.

Traits

Index

Trait object representing a ColumnIndex in Rust’s native format.

Functions

compute_rows

Returns the set of intervals (start, len) containing all the selected rows (for a given column)

select_pages

Given a set of selected Intervals of rows and the set of PageLocation, returns the a set of FilteredPage with the same number of items as locations.