Module polars_parquet::parquet::page

Structs

CompressedDataPage

A CompressedDataPage is compressed, encoded representation of a Parquet data page. It holds actual data and thus cloning it is expensive.

CompressedDictPage

A compressed, encoded dictionary page.

DataPage

A DataPage is an uncompressed, encoded representation of a Parquet data page. It holds actual data and thus cloning it is expensive.

DataPageHeaderV1

Data page header

DataPageHeaderV2

New page format allowing reading levels without decompressing the data Repetition and definition levels are uncompressed The remaining section containing the data is compressed if is_compressed is true

DictPage

An uncompressed, encoded dictionary page.

ParquetPageHeader

Enums

CompressedPage

A CompressedPage is a compressed, encoded representation of a Parquet page. It holds actual data and thus cloning it is expensive.

DataPageHeader

Page

A Page is an uncompressed, encoded representation of a Parquet page. It may hold actual data and thus cloning it may be expensive.

PageType

Traits

DataPageHeaderExt

Functions

split_buffer

Splits the page buffer into 3 slices corresponding to (encoded rep levels, encoded def levels, encoded values).

split_buffer_v1

Splits the page buffer into 3 slices corresponding to (encoded rep levels, encoded def levels, encoded values) for v1 pages.

split_buffer_v2

Splits the page buffer into 3 slices corresponding to (encoded rep levels, encoded def levels, encoded values) for v2 pages.