polars_parquet::parquet::read

Function get_column_iterator

pub fn get_column_iterator<'a>(
    reader: MemReader,
    row_group: &'a RowGroupMetadata,
    field_name: &str,
    max_page_size: usize,
) -> ColumnIterator<'a> 

Returns a ColumnIterator of column chunks corresponding to field.

Contrarily to get_page_iterator that returns a single iterator of pages, this iterator iterates over columns, one by one, and returns a PageReader per column. For primitive fields (e.g. i64), ColumnIterator yields exactly one column. For complex fields, it yields multiple columns. max_page_size is the maximum number of bytes allowed.