polars_parquet::parquet::schema

Module types

Structs

FieldInfo

Common type information.

PrimitiveType

The complete description of a parquet column

Enums

GroupConvertedType

GroupLogicalType

IntegerType

Enum of all valid logical integer types

ParquetType

Representation of a Parquet type describing primitive and nested fields, including the top-level schema of the parquet file.

PhysicalType

The set of all physical types representable in Parquet

PrimitiveConvertedType

PrimitiveLogicalType

TimeUnit