polars_parquet::parquet::write

Function write_metadata_sidecar

pub fn write_metadata_sidecar<W: Write>(
    writer: &mut W,
    metadata: &ThriftFileMetadata,
) -> ParquetResult<u64>

Writes a parquet file containing only the header and footer

This is used to write the metadata as a separate Parquet file, usually when data is partitioned across multiple files.

Note: Recall that when combining row groups from ThriftFileMetadata, the file_path on each of their column chunks must be updated with their path relative to where they are written to.