Function polars_parquet::arrow::read::column_iter_to_arrays

pub fn column_iter_to_arrays<'a, I>(
    columns: Vec<I>,
    types: Vec<&PrimitiveType>,
    field: Field,
    chunk_size: Option<usize>,
    num_rows: usize,
) -> PolarsResult<ArrayIter<'a>>
where
    I: 'a + PagesIter,

An iterator adapter that maps multiple iterators of PagesIter into an iterator of Arrays.

For a non-nested datatypes such as ArrowDataType::Int32, this function requires a single element in columns and types. For nested types, columns must be composed by all parquet columns with associated types types.

The arrays are guaranteed to be at most of size chunk_size and data type field.data_type.