Module polars_parquet::arrow::read::schema

APIs to handle Parquet <-> Arrow schemas.

Re-exports

pub use crate::parquet::metadata::FileMetaData;

pub use crate::parquet::metadata::SchemaDescriptor;

pub use crate::parquet::schema::types::ParquetType;

Structs

KeyValue

Wrapper struct to store key values

SchemaInferenceOptions

Options when inferring schemas from Parquet

Functions

infer_schema

Infers a ArrowSchema from parquet’s FileMetaData. This first looks for the metadata key "ARROW:schema"; if it does not exist, it converts the parquet types declared in the file’s parquet schema to Arrow’s equivalent.

infer_schema_with_options

Like infer_schema but with configurable options which affects the behavior of inference

parquet_to_arrow_schema

Converts ParquetTypes to a Field, ignoring parquet fields that do not contain any physical column.

parquet_to_arrow_schema_with_options

Like parquet_to_arrow_schema but with configurable options which affect the behavior of schema inference

read_schema_from_metadata

Reads an arrow schema from Parquet’s file metadata. Returns None if no schema was found.