polars_parquet::parquet

Module read

Modules

levels

Structs

BasicDecompressor

A [FallibleStreamingIterator] that decompresses CompressedPage into DataPage.

ColumnIterator

A MutStreamingIterator that reads column chunks one by one, returning a PageReader per column.

PageMetaData

This meta is a small part of ColumnChunkMetadata.

PageReader

A fallible Iterator of CompressedDataPage. This iterator reads pages back to back until all pages have been consumed.

ReadColumnIterator

A MutStreamingIterator of pre-read column chunks

Enums

State

State of MutStreamingIterator.

Traits

MutStreamingIterator

A special kind of fallible streaming iterator where advance consumes the iterator.

PageIterator

Functions

decompress

Decompresses the page, using buffer for decompression. If page.buffer.len() == 0, there was no decompression and the buffer was moved. Else, decompression took place.

deserialize_metadata

Parse loaded metadata bytes

get_column_iterator

Returns a ColumnIterator of column chunks corresponding to field.

get_page_iterator

Returns a new PageReader by seeking reader to the beginning of column_chunk.

read_metadata

Reads a FileMetadata from the reader, located at the end of the file.

read_metadata_with_size

Reads a FileMetadata from the reader, located at the end of the file, with known file size.