Module polars_parquet::parquet::write

Structs

ColumnOffsetsMetadata

Compressor

A FallibleStreamingIterator that consumes Page and yields CompressedPage holding a reusable buffer (Vec<u8>) for compression.

DynIter

DynIter is an implementation of a single-threaded, dynamically-typed iterator.

DynStreamingIterator

Dynamically-typed FallibleStreamingIterator.

FileWriter

An interface to write a parquet file. Use start to write the header, write to write a row group, and end to write the footer.

WriteOptions

Write options of different interfaces on this crate

Enums

Version

The parquet version to use

Functions

compress

Compresses an [EncodedPage] into a CompressedPage using compressed_buffer as the intermediary buffer.

write_metadata_sidecar

Writes a parquet file containing only the header and footer

Type Aliases

RowGroupIter